:recycle: Refactoring into individual components
