+++ browser-extension/src/components/instructions/HeaderCard.vue
@@ -1,0 +1,24 @@
+<template>
+  <v-card
+    class="text-center pa-3"
+    elevation="3"
+  >
+    <the-app-icon
+      class="mx-auto"
+      :size="128"
+    />
+    <v-card-title class="pt-0 text-h2 font-weight-bold">AIGELB</v-card-title>
+    <v-card-text class="pb-4">
+      <p>
+        {{ i18n.t("instructions.header.subTitle") }}
+      </p>
+    </v-card-text>
+  </v-card>
+</template>
+
+<script setup lang="ts">
+import { i18n } from "#i18n";
+import { VCard, VCardText, VCardTitle } from "vuetify/components";
+
+import TheAppIcon from "@/components/common/TheAppIcon.vue";
+</script>
